Update MCAS_Activity_Playbook
This branch uses an Azure Storage Table to maintain last run times and improves the For Each function to allow for increased Cloud App Security Payloads over 100 entities. 

Due to the use of additional Azure services, ensure that the Connections are updated to reflect your own subscription and resource group, you will also need to create the required Azure Storage Table prior and if the names have changed, this will also need to be updated, Look for the XXXX's for areas which will need to be updated.

Names used for additional Services
      Azure Storage Table - MCASactivitiesLOGdate
                Table PartitionKey - MCAS
                RowKey - Activity
      [Connection] Azure Log Analytics - azureloganalyticsdatacollector
      [Connection] Azure Storage Table -  azuretables

This LogicApp can be extended to leverage KeyVault to store the API Token, which is highly recommended.